// src/app/interfaces/book.interfaces.ts
export interface IBook {
	title: string;
	author: string;
	genre: "FICTION" | "NON_FICTION" | "SCIENCE" | "HISTORY" | "BIOGRAPHY" | "FANTASY";
	isbn: string;
	description: string;
	copies: number;
	available: boolean;
}

export interface IDeductCopies {
	deductCopies(quantity: number): Promise<void>;
}